Move 1-bit backend under quantization
