+++ mlx_vlm/quantization/__init__.py
@@ -1,0 +1,1 @@
+"""Weight quantization backends."""
